requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
pymongo==4.6.1
python-dotenv==1.0.1
fake-useragent==1.4.0
beautifulsoup4==4.12.3
requests==2.31.0
elasticsearch==8.11.1 